fix: first clanky zoom

+++ src/main/java/de/flojo/jam/game/board/IAmMoveable.java
@@ -1,5 +1,5 @@
 package de.flojo.jam.game.board;
 
 public interface IAmMoveable {
-    void move(int rx, int ry);
+    void move(float rx, float ry);
 }
